# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CRUD управления пользователями с JWT аутентификацией",
        "title": "User Test",
        "contact": {
            "name": "akbota",
            "email": "akbota.bolat678@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список всех зарегистрированных пользователей. Требует JWT аутентификации.\n\n**Пример запроса curl:**\n```bash\ncurl -X GET \"http://localhost:8080/api/users\" -H \"Authorization: Bearer $TOKEN\"\n```",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Пользователи"
                ],
                "summary": "👥 Получить всех пользователей",
                "responses": {
                    "200": {
                        "description": "Список пользователей",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ServerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.User"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Требуется аутентификация",
                        "schema": {
                            "$ref": "#/definitions/models.ServerResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/models.ServerResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создает нового пользователя в системе. Логин должен быть уникальным.\n**Пример запроса curl:**\n```bash\ncurl -X POST \"http://localhost:8080/api/users\" -H \"Content-Type: application/json\" -d '{\n\"first_name\": \"Иван\",\n\"last_name\": \"Петров\",\n\"login\": \"ivan@example.com\",\n\"password\": \"Password123!\"\n}'\n```",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Пользователи"
                ],
                "summary": "📝 Регистрация нового пользователя",
                "parameters": [
                    {
                        "description": "Данные нового пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Пользователь успешно создан",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ServerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Неверный формат данных",
                        "schema": {
                            "$ref": "#/definitions/models.ServerResponse"
                        }
                    },
                    "409": {
                        "description": "Пользователь с таким логином уже существует",
                        "schema": {
                            "$ref": "#/definitions/models.ServerResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/models.ServerResponse"
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "Аутентифицирует пользователя по логину и паролю, возвращает JWT токен",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Аутентификация"
                ],
                "summary": "🔐 Вход в систему",
                "parameters": [
                    {
                        "description": "Учетные данные",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный вход",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ServerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Неверный формат данных",
                        "schema": {
                            "$ref": "#/definitions/models.ServerResponse"
                        }
                    },
                    "401": {
                        "description": "Неверные учетные данные",
                        "schema": {
                            "$ref": "#/definitions/models.ServerResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает информацию о конкретном пользователе. Требует JWT аутентификации.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Пользователи"
                ],
                "summary": "🔍 Получить пользователя по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные пользователя",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ServerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "$ref": "#/definitions/models.ServerResponse"
                        }
                    },
                    "401": {
                        "description": "Требуется аутентификация",
                        "schema": {
                            "$ref": "#/definitions/models.ServerResponse"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "$ref": "#/definitions/models.ServerResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет информацию о пользователе. Все поля опциональны. Требует JWT аутентификации.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Пользователи"
                ],
                "summary": "✏️ Обновить пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь обновлен",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ServerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "schema": {
                            "$ref": "#/definitions/models.ServerResponse"
                        }
                    },
                    "401": {
                        "description": "Требуется аутентификация",
                        "schema": {
                            "$ref": "#/definitions/models.ServerResponse"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "$ref": "#/definitions/models.ServerResponse"
                        }
                    },
                    "409": {
                        "description": "Логин уже занят",
                        "schema": {
                            "$ref": "#/definitions/models.ServerResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет пользователя из системы. Требует JWT аутентификации.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Пользователи"
                ],
                "summary": "🗑️ Удалить пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь удален",
                        "schema": {
                            "$ref": "#/definitions/models.ServerResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "$ref": "#/definitions/models.ServerResponse"
                        }
                    },
                    "401": {
                        "description": "Требуется аутентификация",
                        "schema": {
                            "$ref": "#/definitions/models.ServerResponse"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "$ref": "#/definitions/models.ServerResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CreateUserRequest": {
            "type": "object",
            "required": [
                "first_name",
                "last_name",
                "login",
                "password"
            ],
            "properties": {
                "first_name": {
                    "type": "string",
                    "example": "Дина"
                },
                "last_name": {
                    "type": "string",
                    "example": "Алмазова"
                },
                "login": {
                    "type": "string",
                    "example": "dina@gmail.com"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "models.LoginRequest": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "example": "dina@gmail.com"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "models.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
        "models.ServerResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": "Пользователь с таким логином уже существует"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string",
                    "minLength": 2,
                    "example": "Дана"
                },
                "last_name": {
                    "type": "string",
                    "minLength": 2,
                    "example": "Алмасова"
                },
                "login": {
                    "type": "string",
                    "minLength": 3,
                    "example": "dina1@gmail.com"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "NewPassword456!"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string",
                    "example": "Дина"
                },
                "last_login": {
                    "type": "string",
                    "example": "2026-01-28T12:30:45+05:00"
                },
                "last_name": {
                    "type": "string",
                    "example": "Алмазова"
                },
                "login": {
                    "type": "string",
                    "example": "dina@gmail.com"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите \"Bearer\" затем пробел и ваш JWT токен. Пример: `Bearer eyJhbGciOiJIUzI1NiIs...`",
            "type": "apiKey",
            "name": "just simple project",
            "in": "header"
        }
    }
}